#14 refactor: 디렉토리 구조 변경
