+++ backend/src/main/java/org/example/backend/department/domain/entity/Department.java
@@ -1,9 +1,9 @@
-package org.example.backend.domain.entity;
+package org.example.backend.department.domain.entity;
 
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
-import org.example.backend.domain.dto.Department.DepartmentDto;
+import org.example.backend.department.domain.dto.Department.DepartmentDto;
 
 @Entity
 @Getter
