+++ backend/src/main/java/org/example/backend/department/repository/DepartmentRepository.java
@@ -1,6 +1,6 @@
-package org.example.backend.repository;
+package org.example.backend.department.repository;
 
-import org.example.backend.domain.entity.Department;
+import org.example.backend.department.domain.entity.Department;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DepartmentRepository extends JpaRepository<Department, Long> {
