+++ backend/src/main/java/org/example/backend/professor/controller/ProfessorController.java
@@ -1,11 +1,11 @@
-package org.example.backend.controller;
+package org.example.backend.professor.controller;
 
 
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
-import org.example.backend.domain.dto.professor.ProfessorReqDto;
-import org.example.backend.domain.dto.professor.ProfessorResDto;
-import org.example.backend.service.ProfessorService;
+import org.example.backend.professor.domain.dto.professor.ProfessorReqDto;
+import org.example.backend.professor.domain.dto.professor.ProfessorResDto;
+import org.example.backend.professor.service.ProfessorService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
