+++ backend/src/main/java/org/example/backend/professor/repository/ProfessorRepository.java
@@ -1,6 +1,6 @@
-package org.example.backend.repository;
+package org.example.backend.professor.repository;
 
-import org.example.backend.domain.entity.Professor;
+import org.example.backend.professor.domain.entity.Professor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProfessorRepository extends JpaRepository<Professor, Long> {
